+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/ActionDeserializers/FolderV1Deserializer.cs
@@ -1,0 +1,24 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.ActionDeserializers;
+
+using System;
+using Newtonsoft.Json.Linq;
+using RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data;
+using RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data.Actions;
+
+public class FolderV1Deserializer : IActionDeserializer
+{
+    bool IActionDeserializer.CanDeserialize(string type)
+    {
+        return "folder@1".Equals(type, StringComparison.CurrentCultureIgnoreCase);
+    }
+
+    RepositoryAction IActionDeserializer.Deserialize(JToken jToken)
+    {
+        return Deserialize(jToken);
+    }
+
+    public RepositoryActionFolderV1 Deserialize(JToken jToken)
+    {
+        return jToken.ToObject<RepositoryActionFolderV1>();
+    }
+}
