+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/Data/Actions/RepositoryActionBrowserV1.cs
@@ -1,0 +1,6 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data.Actions;
+
+public class RepositoryActionBrowserV1 : RepositoryAction
+{
+    public string Url { get; set; }
+}
