+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/Data/Actions/RepositoryActionFolderV1.cs
@@ -1,0 +1,8 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data.Actions;
+
+using System.Collections.Generic;
+
+public class RepositoryActionFolderV1 : RepositoryAction
+{
+    public List<RepositoryAction> Actions { get; set; }
+}
