+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/Data/FileReference.cs
@@ -1,0 +1,8 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data;
+
+public class FileReference
+{
+    public string Filename { get; set; }
+
+    public string When { get; set; }
+}
