+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/Data/Redirect.cs
@@ -1,0 +1,8 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data;
+
+public class Redirect
+{
+    public string Filename { get; set; }
+
+    public string Enabled { get; set; }
+}
