+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/Data/RepositoryActionTag.cs
@@ -1,0 +1,8 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data;
+
+public class RepositoryActionTag
+{
+    public string Tag { get; set; }
+
+    public string When { get; set; }
+}
