+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/Data/Variable.cs
@@ -1,0 +1,10 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data;
+
+public class Variable
+{
+    public string Name { get; set; }
+
+    public string Value { get; set; }
+
+    public string Enabled { get; set; }
+}
