+++ src/RepoZ.Api.Common/IO/ModuleBasedRepositoryActionProvider/IActionDeserializer.cs
@@ -1,0 +1,10 @@
+namespace RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider;
+
+using Newtonsoft.Json.Linq;
+using RepoZ.Api.Common.IO.ModuleBasedRepositoryActionProvider.Data;
+
+public interface IActionDeserializer
+{
+    bool CanDeserialize(string type);
+    RepositoryAction Deserialize(JToken jToken);
+}
