+++ src/RepoZ.Api.Common/RepoZ.Api.Common.csproj
@@ -20,8 +20,4 @@
     <ProjectReference Include="..\RepoZ.Api\RepoZ.Api.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="IO\ModuleBasedRepositoryActionProvider\" />
-  </ItemGroup>
-
 </Project>
