+++ tests/RepoZ.Api.Common.Tests/RepoZ.Api.Common.Tests.csproj
@@ -25,8 +25,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\RepoZ.Api.Common\RepoZ.Api.Common.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="IO\ModuleBasedRepositoryActionProvider\" />
-  </ItemGroup>
 </Project>
